django-back/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.12
django-allauth==0.51.0
django-cors-headers==3.12.0
django-environ==0.9.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
idna==3.3
lunardate==0.2.0
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytimekr==0.1.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
